separate (some) CCP code
