+++ keyboards/keychron/k3_pro/ansi/rgb/keymaps/horwitz/layers.h
@@ -1,0 +1,13 @@
+#ifndef LAYERS_H_INCLUDED
+#define LAYERS_H_INCLUDED
+
+enum layers {
+    MAC_BASE,
+    MAC_FN,
+    WIN_BASE,
+    WIN_FN,
+    GCP, // [GCP]
+    CCP // [CCP]
+};
+
+#endif
