+++ keyboards/keychron/k3_pro/ansi/rgb/keymaps/horwitz/rules.mk
@@ -1,2 +1,4 @@
 TAP_DANCE_ENABLE = yes
 CONSOLE_ENABLE = yes
+
+SRC += ccp.c
